# Extend Dictionary concurrent access detection to Remove()