closes https://github.com/dotnet/corefx/issues/30023

I used Ben's idea  https://github.com/dotnet/corefx/commit/6df07c4ecc4d055d47f43202e6b3d5d3e4331374

Do we need some tests?It's not easy to repro...but if you've some fast way to break(maybe some old code used for FindEntry) i could test in local.

/cc @danmosemsft @stephentoub @benaadams 

NB. A few years ago i lost hours to spot this bug(dump/WinDbg) in a third-party lib, 100% cpu on FindEntry.