@MarcoRossignoli 2018-06-18
> If I wanted to test, I would use reflection in the test eg to add a loop in entries.

I could try to add a tests case for CI, what do you think?