@jkotas 2018-06-18
@dotnet-bot test this please