@danmosemsft 2018-06-18
@MarcoRossignoli test would be great, it is a bit hacky to use reflection but we don't want to regress this as its such an impactful issue.would you mind adding the same test for the existing cases? Should be trivial when you have done one.