@MarcoRossignoli 2018-06-18
> would you mind adding the same test for the existing cases? 

Sure!