@MarcoRossignoli 2018-06-18
@danmosemsft do you wait tests to merge this?