@MarcoRossignoli 2018-06-18
> I think if you're able to write them now, there is no reason to merge this without it.

Ok!Working on it!