@Wraith2 2018-06-18
Can I ask why the change from _entries to the local entries? I was looking at implementing RemoveAll and found both direct and local access in dictionary and wasn't sure which was preferable. Clearly local is better in this case but why?