@danmosemsft 2018-06-18
@Wraith2 I understand that it produces better code gen with the current JIT. We only do tricks like that where it's really hot, as it's not a good idea to code around the JIT in general. @AndyAyersMS is this correct, and do you expect the JIT to be able to do this itself in future?