@MarcoRossignoli 2018-06-18
@danmosemsft @Wraith2 my idea is to be more close as possible to code style in other two methods with concurrent access detection [FindEntry](https://github.com/dotnet/coreclr/blob/master/src/System.Private.CoreLib/shared/System/Collections/Generic/Dictionary.cs#L364)/[TryInsert](https://github.com/dotnet/coreclr/blob/master/src/System.Private.CoreLib/shared/System/Collections/Generic/Dictionary.cs#L472)