@Wraith2 2018-06-18
@MarcoRossignoli understood, and it looks like the fix I'd put into my working version. I just wondered at the choice of change since I'm trying to feel my own way around what I should and shouldn't change. 

TryInsert for example mixes a local for entries with direct access to buckets while FindEntry has locals for both, I wouldn't want to introduce performance degradations in something as central as Dictionary by not understanding the choice if there was one. It might be that it matters depending on the number of accesses or where they are etc. This seemed like a good place to ask.