@AndyAyersMS 2018-06-18
In code like this where access to a class member like `entries` is interleaved with other computation, including calls:
```C#
    ... this.entries ...
    ... code that makes calls ...
    ... this.entries ... (probably unmodified from above)
```
the jit is going to assume that the this `this.entries` member can be reassigned by the calls and so the jit will likely reload it at each use. Caching the member in a local (when valid) makes it more obvious the value does not change. The jit will then try to keep the value in a register and can more readily propagate facts about the value (like non-nullness or bounds check state) that can enable other optimizations.

So generally speaking I would expect caching a frequently accessed member in a local would be a good idea perf wise, though it is very important to measure and be sure. Sometimes the jit has a lot of competition for things that can go in registers and adding one more "long lifetime" resource to the mix can have unexpected consequences.

Mixing locally cached and direct member access in one method is confusing (though sometimes necessary if the member value can indeed change). But I'd avoid mixing usage unless it's really needed for correctness, and if it really is necessary, make sure to comment on this in detail.

Unfortunately there's no easy way to catch if you mix usages by accident other than careful review of the source and perhaps the generated assembly.