@MarcoRossignoli 2018-06-18
@AndyAyersMS thank's a lot for explanation! So in this case cache values could improve perf because remove methods does not change entry array or rather when change we set some fields on a particular entry and return(we found entry to "remove"). If i understood well generally speaking coreclr will try to optimize "read-only loop code"...like `while(true)...` search.