@AndyAyersMS 2018-06-18
For `Dictionary<K,T>` methods the code the jit generates will depend a lot on `K` and `T`, the target ISA and ABI, and whether the code is prejitted or jitted. It will not depend on much else.

For some popular combinations of types the code for the methods will be prejitted ahead of time via crossgen. For other types the code will be jitted the first time the method is called.

To view the code the jit generates you can do a number of things:
* Run the app under a debugger and break in when the method is called and disassemble. Make sure debugger does not disable jit optimizations
* Run via BenchmarkDotNet and use the disassembly diagnoser
* Build a Checked local CoreCLR and enable jit disassembly (by setting COMPlus_JitDisasm). See for instance [Viewing Jit Dumps](https://github.com/dotnet/coreclr/blob/master/Documentation/building/viewing-jit-dumps.md)
* Disassemble during crossgen of System.Private.CoreLib  (by setting COMPlus_NgenDisasm)