@danmosemsft 2018-06-19
@MarcoRossignoli it would be interesting to run the perf tests for dictionary before and after your change. They might also suggest the effect of extracting locals in this way.

They are [here](https://github.com/dotnet/corefx/blob/master/src/System.Collections/tests/Performance/Perf.Dictionary.cs) -- I see only one Remove test, `Remove_ValueType` -- that's probably OK. You could add a similar one (even if you don't check it in) for the other overload.

https://github.com/dotnet/corefx/blob/master/Documentation/project-docs/performance-tests.md explains running corefx perf tests - you already know how to patch CoreFX with CoreCLR - this time it would of course be release built bits.