@MarcoRossignoli 2018-06-19
@danmosemsft  if i compile with my local clr bits(updated cleaned repos) i get compile error:
```
MembersMustExist : Member 'System.Runtime.Intrinsics.X86.Avx.Extract(System.Runtime.Intrinsics.Vector256<System.Int16>, System.Byte)'
...
```
all errors are on `System.Runtime.Intrinsics.Experimental`, is this expected?
This is not first time...other times worked well.