@jkotas 2018-06-19
This happens as breaking changes propagate through the system. This one will fix itself up once https://github.com/dotnet/corefx/pull/30497 is merge. You can ignore this. 

To avoid this sort of build break, you would need to sync your coreclr repo back to the build that corefx is using: find the CoreCLR package version in corefx/dependencies.props and then find the hash that the package build from at https://dotnet.myget.org/feed/dotnet-core/package/nuget/runtime.osx-x64.Microsoft.NETCore.Runtime.CoreCLR. We plan to have a better system for this.