@MarcoRossignoli 2018-06-19
> This happens as breaking changes propagate through the system. This one will fix itself up once dotnet/corefx#30497 is merge. You can ignore this.

Ok i'll wait merge.

>To avoid this sort of build break, you would need to sync your coreclr repo back to the build that corefx is using: find the CoreCLR package version in corefx/dependencies.props and then find the hash that the package was built from at https://dotnet.myget.org/feed/dotnet-core/package/nuget/runtime.osx-x64.Microsoft.NETCore.Runtime.CoreCLR. We plan to have a better system for this.

Understood..it means i need to wait to "test my local updates"...because on myget i can download only compiled package...no sources...isn't it?Or there is a way to `sync your coreclr repo back` at source level?

my props file(working on win):
```
<CoreClrCurrentRef>9a1f1f4afe718228ba1252b43a2cebb7165bacd1</CoreClrCurrentRef>
...
<MicrosoftNETCoreRuntimeCoreCLRPackageVersion>2.2.0-preview1-26616-03</MicrosoftNETCoreRuntimeCoreCLRPackageVersion>
```
As i said i can wait...but it would be great be able to "work" also in this case(until better way), i don't want to waste precious spare time :smiley:.