@MarcoRossignoli 2018-06-19
Tests code: 
```cs
        [Benchmark]
        [InlineData(20000)]
        [InlineData(100000)]
        [InlineData(1000000)]
        public static void Remove_ValueType(long size)
        {
            Dictionary<long?, long?> collection = new Dictionary<long?, long?>();
            long?[] items;

            items = new long?[size * 10];

            for (long i = 0; i < size * 10; ++i)
            {
                items[i] = i;
                collection.Add(items[i], items[i]);
            }


            foreach (var iteration in Benchmark.Iterations)
                using (iteration.StartMeasurement())
                    for (long i = 1; i < size; ++i)
                        collection.Remove(items[i]);
        }

        [Benchmark]
        [InlineData(20000)]
        [InlineData(100000)]
        [InlineData(1000000)]
        public static void Remove_ValueType_Out(long size)
        {
            Dictionary<long?, long?> collection = new Dictionary<long?, long?>();
            long?[] items;

            items = new long?[size * 10];

            for (long i = 0; i < size * 10; ++i)
            {
                items[i] = i;
                collection.Add(items[i], items[i]);
            }


            foreach (var iteration in Benchmark.Iterations)
                using (iteration.StartMeasurement())
                    for (long i = 1; i < size; ++i)
                        collection.Remove(items[i], out long? value);
        }
```

Before 

 System.Collections.Performance.Tests.dll                                     | Metric   | Unit | Iterations | Average | STDEV.S |   Min |    Max
:---------------------------------------------------------------------------- |:-------- |:----:|:----------:| -------:| -------:| -----:| ------:
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType(size: 100000)      | Duration | msec |    1000    |   0.800 |   0.225 | 0.707 |  2.285
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType(size: 1000000)     | Duration | msec |    1000    |   7.662 |   0.245 | 7.388 | 13.435
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType(size: 20000)       | Duration | msec |    1000    |   0.152 |   0.014 | 0.142 |  0.422
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType_Out(size: 100000)  | Duration | msec |    1000    |   0.894 |   0.063 | 0.828 |  1.264
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType_Out(size: 1000000) | Duration | msec |    1000    |   8.994 |   0.203 | 8.704 | 10.738
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType_Out(size: 20000)   | Duration | msec |    1000    |   0.179 |   0.018 | 0.166 |  0.347

After

 System.Collections.Performance.Tests.dll                                     | Metric   | Unit | Iterations | Average | STDEV.S |   Min |    Max
:---------------------------------------------------------------------------- |:-------- |:----:|:----------:| -------:| -------:| -----:| ------:
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType(size: 100000)      | Duration | msec |    1000    |   0.796 |   0.065 | 0.733 |  1.391
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType(size: 1000000)     | Duration | msec |    1000    |   7.965 |   0.213 | 7.641 | 11.305
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType(size: 20000)       | Duration | msec |    1000    |   0.160 |   0.020 | 0.147 |  0.341
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType_Out(size: 100000)  | Duration | msec |    1000    |   0.877 |   0.056 | 0.816 |  1.260
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType_Out(size: 1000000) | Duration | msec |    1000    |   8.843 |   0.551 | 8.532 | 16.917
 System.Collections.Tests.Perf_Dictionary.Remove_ValueType_Out(size: 20000)   | Duration | msec |    1000    |   0.182 |   0.028 | 0.164 |  0.572


@AndyAyersMS @danmosemsft these are perf test result...seems there is no great difference, let me know if code test code is correct(first time with  "Microsoft xunit-performance framework".

What i've done:

1) checkout master build release
2) build corefx in release with my local clr bits
3) buildandtest perf 3 times...checked that values are all similar
4) checkout my clr branch with updates(this PR) 
5) build -skiptests -release -skipnative
6) buildandtest perf 3 times...checked that values are all similar

Done some times.
Let me know if results make sense for you.